+++ tests/test_cache.py
@@ -1,4 +1,6 @@
 
+import re
+
 import pytest
 
 from cacheout import Cache
@@ -128,6 +130,29 @@ def test_cache_get_many(cache):
     assert cache.get_many(items.keys()) == items
 
 
+@parametrize('items,iteratee,expected', [
+    ({'a_1': 1, 'a_2': 2, 'bcd': 3, 'bed': 4, '12345': 5},
+     ['a_1', '12345'],
+     {'a_1': 1, '12345': 5}),
+    ({'a_1': 1, 'a_2': 2, 'bcd': 3, 'bed': 4, '12345': 5},
+     'a_*',
+     {'a_1': 1, 'a_2': 2}),
+    ({'a_1': 1, 'a_2': 2, 'bcd': 3, 'bed': 4, '12345': 5},
+     re.compile(r'\d'),
+     {'12345': 5}),
+    ({'a_1': 1, 'a_2': 2, 'bcd': 3, 'bed': 4, '12345': 5},
+     lambda key: key.startswith('b') and key.endswith('d'),
+     {'bcd': 3, 'bed': 4}),
+])
+def test_cache_get_many_by(cache, items, iteratee, expected):
+    """Test that cache.get_many_by() returns multiple cache key/values filtered
+    by an iteratee.
+    """
+    cache.set_many(items)
+
+    assert cache.get_many_by(iteratee) == expected
+
+
 def test_cache_delete(cache):
     """Test that cache.delete() removes a cache key."""
     key, value = ('key', 'value')
@@ -151,6 +176,31 @@ def test_cache_delete_many(cache):
     cache.delete_many(items.keys())
 
     assert len(cache) == 0
+
+
+@parametrize('items,iteratee,expected', [
+    ({'a_1': 1, 'a_2': 2, 'bcd': 3, 'bed': 4, '12345': 5},
+     ['a_1', '12345'],
+     {'a_2': 2, 'bcd': 3, 'bed': 4}),
+    ({'a_1': 1, 'a_2': 2, 'bcd': 3, 'bed': 4, '12345': 5},
+     'a_*',
+     {'bcd': 3, 'bed': 4, '12345': 5}),
+    ({'a_1': 1, 'a_2': 2, 'bcd': 3, 'bed': 4, '12345': 5},
+     re.compile(r'\d'),
+     {'a_1': 1, 'a_2': 2, 'bcd': 3, 'bed': 4}),
+    ({'a_1': 1, 'a_2': 2, 'bcd': 3, 'bed': 4, '12345': 5},
+     lambda key: key.startswith('b') and key.endswith('d'),
+     {'a_1': 1, 'a_2': 2, '12345': 5}),
+])
+def test_cache_delete_many_by(cache, items, iteratee, expected):
+    """Test that cache.delete_many_by() deletes multiple cache key/values
+    filtered by an iteratee.
+    """
+    cache.set_many(items)
+
+    cache.delete_many_by(iteratee)
+
+    assert dict(cache.items()) == expected
 
 
 def test_cache_delete_expired(cache, timer):